requirements.txt
streamlit
langchain
langchain-community
langchain_text_splitters
langchain_huggingface
langchain_groq
python-dotenv
pypdf
faiss-cpu